% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carithmetic.R
\name{binary_string_to_rational}
\alias{binary_string_to_rational}
\title{Convert a binary string to its rational representation}
\usage{
binary_string_to_rational(binary_string)
}
\arguments{
\item{binary_string}{A string representing a binary number, which may include a fractional part.}
}
\value{
A list containing the floating-point approximation, numerator, and denominator of the rational number.
}
\description{
Convert a binary string to its rational representation
}
