% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{columns_to_factors}
\alias{columns_to_factors}
\title{Convert columns to factors.}
\usage{
columns_to_factors(data, col.names.use = NULL)
}
\arguments{
\item{data}{the data frame containing the (two or more) numerical columns.}

\item{col.names.use}{a vector containing the desired column names as strings.}
}
\description{
Function to convert a data frame with two or more columns of numerical values, where each column corresponds to a specific factor level, into two columns, where the first column contains all of the numerical values and the second column contains their corresponding factor level.
}
