% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carithmetic.R
\name{convert.decimal.fraction.to.nary}
\alias{convert.decimal.fraction.to.nary}
\title{Convert a decimal fraction into its base-n (n-nary) representation using
integer arithmetic with Horner's method.}
\usage{
convert.decimal.fraction.to.nary(d, e, base, ndigits = 54)
}
\arguments{
\item{d}{the integer part of the decimal fraction.}

\item{e}{the exponent for 10^(-e)}

\item{base}{the base for the n-ary representation of the decimal fraction}

\item{ndigits}{the maximum number of digits to search for the repeating pattern}
}
\description{
A function to convert a decimal fraction into its base-n
(n-ary) representation.
}
\details{
See

http://cs.furman.edu/digitaldomain/more/ch6/dec_frac_to_bin.htm

for more details.

**NOTE:** d, e, and base should be **integer type**, i.e.
suffixed with an `L`.

If not, they will be coerced to integer type if possible.
}
\examples{
# Binary expansion of 1/10 = 1/10^1

convert.decimal.fraction.to.nary(1, 1, base = 2)

# Trinary expansion of 1/10 = 1/10^1

convert.decimal.fraction.to.nary(5, 1, base = 3)

# Base-5 expansion of 1/5 = 2/10^1

convert.decimal.fraction.to.nary(2, 1, base = 5)
}
