% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carithmetic.R
\name{convert.rational.fraction.to.nary}
\alias{convert.rational.fraction.to.nary}
\title{Convert a rational fraction into its base-n (n-nary) representation using
integer arithmetic with Horner's method.}
\usage{
convert.rational.fraction.to.nary(p, q, base, ndigits = 1000)
}
\arguments{
\item{p}{the numerator of the rational fraction}

\item{q}{the denominator of the rational fraction}

\item{base}{the base for the n-ary representation of the decimal fraction}

\item{ndigits}{the maximum number of digits to search for the repeating pattern}
}
\description{
A function to convert a rational fraction into its base-n
(n-ary) representation.
}
\details{
See

http://cs.furman.edu/digitaldomain/more/ch6/dec_frac_to_bin.htm

for more details.

**NOTE:** p, q, and base should be **integer type**, i.e.
suffixed with an `L`.

If not, they will be coerced to integer type if possible.
}
\examples{
# Binary expansion of 1/10

convert.rational.fraction.to.nary(1, 10, base = 2)

# Base-5 expansion of 1/3

convert.decimal.fraction.to.nary(1, 3, base = 5)
}
