% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carithmetic.R
\name{first.decimal.digit}
\alias{first.decimal.digit}
\title{Get out the leading (left-most) digit in the base-10
representation of an integer / float.}
\usage{
first.decimal.digit(number)
}
\arguments{
\item{number}{an integer / float}
}
\description{
Get out the leading (left-most) digit in the base-10
representation of an integer / float.
}
