% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{frequency_table_from_gf_histogram}
\alias{frequency_table_from_gf_histogram}
\title{Generate a frequency table from a ggformula histogram object.}
\usage{
frequency_table_from_gf_histogram(object, plot = TRUE)
}
\arguments{
\item{object}{An output from gf_histogram.}

\item{plot}{Logical indicating whether (TRUE) or not (FALSE) to plot the histogram passed to frequency_table_from_gf_histogram.}
}
\description{
Function to create a frequency table from a ggformula histogram object. Returns the cut points, bins, and counts within each bin.
}
