% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carithmetic.R
\name{integer.to.integer.power}
\alias{integer.to.integer.power}
\title{Raise an integer to an integer power}
\usage{
integer.to.integer.power(base, exponent)
}
\arguments{
\item{base}{the integer base of the exponential}

\item{exponent}{the integer exponent of the exponential}
}
\description{
Raise a fixed-point integer to a fixed-point integer power,
while maintaining the numeric type of the integer, returning
}
\details{
base^exponent

This is needed since in R, a^b, with a and b both integer,
returns a float rather than an integer.
}
