% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{one.sample.t.test}
\alias{one.sample.t.test}
\title{Gosset's one-sample t-test with summary statistics}
\usage{
one.sample.t.test(xbar, s, n, mu0 = 0, alternative = c("two.sided",
  "less", "greater"), conf.level = 0.95)
}
\arguments{
\item{xbar}{the sample mean.}

\item{s}{the sample standard deviation.}

\item{n}{the sample size.}

\item{mu0}{the assumed population mean under the null hypothesis.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{conf.level}{confidence level for the interval estimator}
}
\description{
Gosset's one-sample t-test for testing claims about a population mean.
}
