% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{two.sample.t.test}
\alias{two.sample.t.test}
\title{Welch's two-sample t-test with summary statistics}
\usage{
two.sample.t.test(xbar, ybar, sx, sy, nx, ny, null.diff = 0,
  alternative = c("two.sided", "less", "greater"), conf.level = 0.95)
}
\arguments{
\item{xbar}{the sample mean of the first sample.}

\item{ybar}{the sample mean of the second sample.}

\item{sx}{the sample standard deviation of the first sample.}

\item{sy}{the sample standard deviation of the second sample.}

\item{nx}{the sample size of the first sample.}

\item{ny}{the sample size of the second sample.}

\item{null.diff}{the assumed difference mu_X - mu_Y under the null hypothesis.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{conf.level}{confidence level for the interval estimator}
}
\description{
Welch's two-sample t-test for testing claims about the different between two population means without assuming the population variances are equal.
}
